+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## master
+## 0.4.18 / 2013-12-17
 
 * [improvement] `shelly add` should ask about databases instead of database
 * [improvement] Improved messages displayed while deleting the application